Added confirmation messages for deletions and handling for when the server is down

+++ client/src/main/java/client/scenes/AdminEventOverviewCtrl.java
@@ -8,10 +8,7 @@ import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.event.EventHandler;
 import javafx.fxml.FXML;
-import javafx.scene.control.Alert;
-import javafx.scene.control.Button;
-import javafx.scene.control.TableColumn;
-import javafx.scene.control.TableView;
+import javafx.scene.control.*;
 import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.scene.input.KeyCode;
 import javafx.scene.input.MouseEvent;
@@ -83,10 +80,16 @@ public class AdminEventOverviewCtrl {
             Button deleteButton = new Button("Delete");
             deleteButton.setOnAction(event -> {
                 Event selectedEvent = param.getValue();
-                server.deleteEvent(selectedEvent.getId());
-                events.remove(selectedEvent);
+                var alert = new Alert(Alert.AlertType.CONFIRMATION);
+                alert.initModality(Modality.APPLICATION_MODAL);
+                alert.setContentText("Are you sure you want to delete this event?");
+                alert.showAndWait().ifPresent((response)->{
+                    if(response == ButtonType.OK){
+                        server.deleteEvent(selectedEvent.getId());
+                        events.remove(selectedEvent);
+                    }
+                });
                 refresh();
-                popup("event deleted");
             });
             return new SimpleObjectProperty<>(deleteButton);
         });
@@ -160,7 +163,11 @@ public class AdminEventOverviewCtrl {
         var alert = new Alert(Alert.AlertType.CONFIRMATION);
         alert.initModality(Modality.APPLICATION_MODAL);
         alert.setContentText(message);
-        alert.showAndWait();
+        alert.showAndWait().ifPresent((response)->{
+            if(response == ButtonType.OK){
+
+            }
+        });
     }
 
     /**
