+++ client/src/main/java/client/scenes/OverviewCtrl.java
@@ -28,6 +28,7 @@ import javafx.scene.paint.Color;
 import javafx.scene.text.Text;
 import commons.Expense;
 import javafx.collections.ObservableList;
+import javafx.stage.Modality;
 
 
 import java.io.*;
@@ -216,25 +217,33 @@ public class OverviewCtrl implements Main.UpdatableUI {
             label.setOnMouseClicked(mouseEvent -> {
                 if (mouseEvent.getClickCount() == 2) {
                     if (mouseEvent.getButton() == MouseButton.SECONDARY) {
-                        event.removeParticipant(contact);
-                        serverUtils.updateEvent(event);
-                        List<Expense> toDelete = new ArrayList<>();
-                        for(Expense expense : event.getExpenses()){
-                            if(expense.getPaidBy().equals(contact)) toDelete.add(expense);
-                            else if(expense.getInvolvedParticipants().contains(contact)){
-                                if(expense.getInvolvedParticipants().size()==1) toDelete.add(expense);
-                                else{
-                                    expense.getInvolvedParticipants().remove(contact);
-                                    serverUtils.updateExpense(event.getId(), expense);
+                        var alert = new Alert(Alert.AlertType.CONFIRMATION);
+                        alert.initModality(Modality.APPLICATION_MODAL);
+                        alert.setContentText("Are you sure you want to delete this participant?");
+                        alert.showAndWait().ifPresent((response)->{
+                            if(response == ButtonType.OK){
+                                event.removeParticipant(contact);
+                                serverUtils.updateEvent(event);
+                                List<Expense> toDelete = new ArrayList<>();
+                                for(Expense expense : event.getExpenses()){
+                                    if(expense.getPaidBy().equals(contact)) toDelete.add(expense);
+                                    else if(expense.getInvolvedParticipants().contains(contact)){
+                                        if(expense.getInvolvedParticipants().size()==1) toDelete.add(expense);
+                                        else{
+                                            expense.getInvolvedParticipants().remove(contact);
+                                            serverUtils.updateExpense(event.getId(), expense);
+                                        }
+                                    }
                                 }
+                                for(Expense expense1: toDelete){
+                                    serverUtils.deleteExpense(event.getId(), expense1);
+                                }
+                                serverUtils.deleteParticipant(contact);
+                                this.event = serverUtils.updateEvent(this.event);
+                                participantsDisplay();
                             }
-                        }
-                        for(Expense expense1: toDelete){
-                            serverUtils.deleteExpense(event.getId(), expense1);
-                        }
-                        serverUtils.deleteParticipant(contact);
-                        this.event = serverUtils.updateEvent(this.event);
-                        participantsDisplay();
+                        });
+
                     } else
                         addParticipant1(contact);
                 }
